refactor: http client to reusable module
